refactor(credentials-catalogue): move getStatusL10nIds to common and add to CredentialsCatalogueApi
